refactor(matcher): improve capture handling and sequence matching logic
Refactor Captures struct to store source reference and remove redundant
source arguments in capture methods. Implement SequenceMatcher struct to
handle patterns with multiple metavariables more cleanly. Anchor empty
multiple metavariable matches to appropriate byte ranges for accurate
text spans. Improve test coverage for multiple and empty metavariable captures.

Additionally, reorganize BDD behaviour tests into a module and move scenario
bindings to a dedicated scenarios file for better structure. Rename some
end-to-end tests for clarity and move snapshots to a separate module.

Minor improvements include enhanced error context extraction, and small
documentation additions to the matcher and BDD test modules.
