+++ crates/weaver-syntax/src/tests/behaviour/scenarios.rs
@@ -1,0 +1,113 @@
+//! Scenario bindings for the `weaver-syntax` BDD feature file.
+//!
+//! These functions bind Gherkin scenario names to the step definitions in the
+//! parent module.
+
+use std::cell::RefCell;
+
+use rstest::fixture;
+use rstest_bdd_macros::scenario;
+
+use super::TestWorld;
+
+/// Fixture providing the shared BDD world.
+#[fixture]
+fn world() -> RefCell<TestWorld> {
+    super::world()
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Valid Rust code passes syntactic validation"
+)]
+fn valid_rust_validation(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Invalid Rust code fails with error location"
+)]
+fn invalid_rust_validation(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Valid Python code passes syntactic validation"
+)]
+fn valid_python_validation(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Invalid Python code fails with error location"
+)]
+fn invalid_python_validation(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Invalid TypeScript code fails with error location"
+)]
+fn invalid_typescript_validation(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Unknown file extensions are skipped"
+)]
+fn unknown_extension_skipped(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Multiple files validated together"
+)]
+fn multiple_files_validation(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Pattern matches function definitions"
+)]
+fn pattern_matches_functions(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Pattern captures metavariable values"
+)]
+fn pattern_captures_metavars(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Pattern with no matches returns empty"
+)]
+fn pattern_no_matches(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Rewrite transforms matching code"
+)]
+fn rewrite_transforms_code(world: RefCell<TestWorld>) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/weaver_syntax.feature",
+    name = "Rewrite with no matches leaves code unchanged"
+)]
+fn rewrite_no_changes(world: RefCell<TestWorld>) {
+    drop(world);
+}
